Broken into the modules
 - separated the one big clunky python executable into the standalone modules
